Reformat WCVP input file
In November 2022 the WCVP export changed to fit a publication
purposes by the IPNI team.
The new export does not include a parent POWO ID, that is required when
running the script. Nick Black agreed to produce a dump similar to
the new version, but including the parent POWO ID.
The major reason for the dump change was that it needed to include
all accepted names, regardless having an IPNI ID assigned or not.
There are 3 types of IDs in the dump: the internal WCVP ID,
the POWO ID, that links to POWO's website, and the IPNI ID.
For the names having an IPNI ID, this is the same as POWO ID.
For those records not yet having an IPNI ID, a temporary POWO ID
has been created ending in "-4". This ID links to the website, but
it will change when the IPNI ID is assigned.

For this script's purposes, the POWO ID is renamed in the input file
as Kew ID, that was the nomenclature used in previous exports,
and that it matched 100% the IPNI IDs.

As a quick fix, I have added the parent accepted name and authors
as new columns with dummy data, as they are not used to resolve
the taxonomy, but the column name is used when parsing data in
the dataframe structure.